// src/model/interfaces/reportes/seccionDetalle.ts
export default interface SeccionDetalle {
    per_Id: string
    asi_Id: string
    docente: string
    anioActual: string
    semestreActual: string
    seccion: string
    nivel: string
    planEstudio: string
    asignatura: string
    sede: string
    carrera: string
    facultad: string
    car_Id: string
    sed_Id: string
}